requirements.txt
alembic==1.9.4
anyascii==0.3.2
apparmor==3.1.6
application-utility==1.3.2
argcomplete==2.0.0
asgiref==3.7.2
attrs==22.2.0
autocommand==2.2.2
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.5
Brotli==1.0.9
btrfsutil==6.3.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
click==8.1.5
configobj==5.0.8
cryptography==41.0.2
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.3.2
decorator==5.1.1
distro==1.8.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
drf-spectacular==0.26.4
duplicity==1.2.3
email-validator==2.0.0.post2
ewmh==0.1.6
fasteners==0.18
fastjsonschema==2.18.0
filelock==3.12.0
fixtures==4.1.0
Flask==2.2.5
flask-babel==3.1.0
Flask-BabelEx==0.9.4
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
future==0.18.3
glad2==2.0.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
gssapi==1.8.2
gufw==22.4.0
httplib2==0.22.0
idna==3.4
importlib-metadata==5.0.0
inflect==6.1.0
inflection==0.5.1
itsdangerous==2.1.2
jaraco.context==4.3.0
jaraco.functools==3.8.0
jaraco.text==3.11.1
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
layoutswitcherlib==0.8.36
ldap3==2.9
LibAppArmor==3.1.6
lit==15.0.7.dev0
lxml==4.9.2
Mako==1.2.4
manjaro-sdk==0.1
MarkupSafe==2.1.3
material-color-utilities-python==0.1.5
meson==1.2.0
more-itertools==9.1.0
netsnmp-python==1.0a1
nftables==0.1
npyscreen==4.10.5
oauth2client==4.1.3
ordered-set==4.1.0
packaging==23.1
pacman-mirrors==4.23.2
paramiko==2.11.1
passlib==1.7.4
pbr==5.11.1
Pillow==10.0.0
pipx==1.2.0
platformdirs==3.9.1
ply==3.11
protobuf==4.23.4
psutil==5.9.5
psycopg2==2.9.6
pulsectl==23.5.2
pwquality==1.4.5
pyaml==23.5.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.2
pydantic==1.10.9
PyDrive2==1.16.1
PyGObject==3.44.1
PyJWT==2.8.0
PyNaCl==1.4.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySide6==6.5.1.1
python-dateutil==2.8.2
python-editor==1.0.4
python-xlib==0.33
pytz==2023.3
PyYAML==6.0
ranger-fm==1.9.3
referencing==0.30.0
regex==2023.6.3
reportlab==3.6.12
requests==2.28.2
requests-file==1.5.1
rpds-py==0.9.2
rsa==4.9
setproctitle==1.3.2
shiboken6==6.5.1.1
shiboken6-generator==6.5.1.1
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
speaklater==1.3
SQLAlchemy==1.3.24
sqlparse==0.4.3
sshtunnel==0.4.0
svglib==1.5.1
systemd-python==235
testtools==2.6.0
tinycss2==1.2.1
tldextract==3.4.4
tomli==2.0.1
tqdm==4.65.0
trove-classifiers==2023.7.8
typing_extensions==4.7.1
ueberzug==18.2.1
ufw==0.36.2
uritemplate==4.1.1
urllib3==1.26.15
userpath==1.8.0
validate==5.0.8
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
webencodings==0.5.1
Werkzeug==2.3.2
WTForms==3.0.1
Yapsy==1.12.2
zipp==3.16.0
